requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.6
django-filter==1.1.0
djangorestframework==3.7.1
idna==2.6
pkg-resources==0.0.0
pytz==2017.2
requests==2.18.4
urllib3==1.22
